requirements.txt
blinker==1.4
Flask==0.10.1
Flask-KVSession==0.6.2
Flask-Mail==0.9.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
python-firebase==1.2
redis==2.10.5
requests==2.9.1
simplekv==0.10.0
six==1.10.0
Werkzeug==0.10.1
wheel==0.24.0
